Drop `ignorecase` (and `smartcase`)
The `ignorecase` search setting is definately useful, but when combined
with `smartcase` it can give the illusion of case sensitive search in
some scenarios. It is particulary problematic for `%s/Foo/bar/`
operations becuase when you want `Foo` to change but not `foo`, then
it is left to your smart brain to remember this issue.

An option to `ignorecase` only on `/` search but not on `%s` would be
great.